Add ability to disable discovery
This adds a third option "None" to the AP discovery methods, which will
disable both sending and receiving of UDP discovery packets so that the
AP no longer exchanges information with other APs.

This also shows a message above the AP list to remind the user when AP
discovery is disabled.

This does not remove external tags from the tag DB.